requirements.txt
pytest==9.0.3
pytest-rerunfailures==16.1
python-dotenv==1.2.2
requests==2.33.1
urllib3==2.6.3
